refactor(nuxt): move landing to pages/index.vue and provide period at app root
